requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlmodel==0.0.18
pydantic-settings==2.2.1
anthropic==0.25.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-api-python-client==2.127.0
notion-client==2.2.1
schedule==1.2.1
python-dotenv==1.0.1
